+++ cmd/gen/workflows/flag.go
@@ -11,7 +11,6 @@ import (
 )
 
 const (
-	flagCheckSecrets                   = "check-secrets"
 	flagEnableFloatingMajorVersionTags = "enable-floating-major-tags"
 	flagFlavour                        = "flavour"
 	flagLanguage                       = "language"
@@ -21,7 +20,6 @@ const (
 )
 
 type flag struct {
-	CheckSecrets                   bool
 	EnableFloatingMajorVersionTags bool
 	Flavours                       gen.FlavourSlice
 	Language                       string
@@ -31,7 +29,6 @@ type flag struct {
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
-	cmd.Flags().BoolVar(&f.CheckSecrets, flagCheckSecrets, true, "If true, also generate a secret-scanning workflow. Possible values: true (default), false.")
 	cmd.Flags().BoolVar(&f.EnableFloatingMajorVersionTags, flagEnableFloatingMajorVersionTags, false, "If true, also generate steps and workflows to ensure floating major version tags like \"v1\" after the release creation.")
 	cmd.Flags().VarP(gen.NewFlavourSliceFlagValue(&f.Flavours, gen.FlavourSlice{}), flagFlavour, "f", fmt.Sprintf(`The type of project that you want to generate the workflows for. Possible values: <%s>`, strings.Join(gen.AllFlavours(), "|")))
 	cmd.Flags().StringVarP(&f.Language, flagLanguage, "l", "", "Language of the repo, for generating additional language-specific workflows, like vulnerability remediation.")